+++ cmd/foreverdubbed/main.go
@@ -19,7 +19,7 @@ import (
 	"foreverdubbed/internal/speech"
 )
 
-const version = "0.3.1"
+const version = "0.3.2"
 
 func main() {
 	log.SetFlags(log.Ltime)